08.12 - final refactorings (step 7 - refactor the user controller) - move the user specific exceptions and the UserDetailsServiceImpl to the users package
