+++ src/main/java/com/jparkkennaby/store/users/DuplicateUserException.java
@@ -1,5 +1,5 @@
-package com.jparkkennaby.store.exceptions;
+package com.jparkkennaby.store.users;
 
 public class DuplicateUserException extends RuntimeException {
-    
+
 }
