+++ src/main/java/com/jparkkennaby/store/users/UserNotFoundException.java
@@ -1,5 +1,5 @@
-package com.jparkkennaby.store.exceptions;
+package com.jparkkennaby.store.users;
 
 public class UserNotFoundException extends RuntimeException {
-    
+
 }
